nimble/phy: Fix receiver configuration on late RX
If we are late when scheduling RX, receiver is enabled anyway since we
may just want to scan for "anything". However, current code only enables
receiver but does not configure it so the results seem unpredictable.

This patch fixes this by ensuring that radio is properly configured for
RX in both cases.

X-Original-Commit: 068c09a8da97bc62eddd5084fa1e40f910abf7e8