nimble/mesh: Add possibility to use extended advertising
With this patch Mesh can make use of multi instances which came with
advertising extension. It creates one advertising instance for
PB-ADV and one for PB-GATT and PROXY.

That said, it is possible to create additional instances which are
used for non mesh purpose. Note that this brings limitation on Mesh
Friendship due to timings.

To make use of it application needs to set BLE_EXT_ADV and
BLE_MULTI_ADV_INSTANCES at least to 1. Mesh will use last available
advertising instance for PB-ADV and second to last for PB-GATT and
PROXY