nimble/phy: Allow to selectively enable workaround for nRF52840
There are known anomalies in nRF52840 engineering and final revisions
which, according to Nordic DevZone, should be only applied to chip
version affected. For this reason We need to allow workarounds for #164
and #191 to be enabled selectively.

Default configuration applies to nRF52840 Rev 1 (final silicon) and
Engineering B and C.

When building for nRF52840 Engineering A (found on early versions on
nRF52840 PDK) configuration should be changed.