nimble/ll: Fix assertion when scheduling aux
Advertising events are scheduled from LL task and we always try to
schedule two auxes one after another (for chaining). However, since this
does not happen in critical section it can happen that we're interrupted
after scheduling 1st aux, so when we try to schedule 2nd one there is no
aux scheduled anymore which causes an assert.

It it safe to just skip the scheduling in this case and thus handle this
gracefully. The only drawback is that we possibly truncate aux chain,
but this is harmless.