nimble/ll: Refactor advertising duration handling
The advertising duration was not handled properly in some cases when
extended advertising was used which resulted in very weird behaviour
over the air (like bursts of events which did not follow interval) or
simply exceeding requested duration.

This fixes all known problems by ensuring that:
- if 1st aux scheduled in chain does not fit within duration, we stop
  advertising immediately
- if subsequent aux scheduled in chain does not fit within duration,
  we remove it from scheduler (data will be trimmed)
- after aux chain is done, we check if next advertising fits within
  duration before scheduling (prior to this we checked if current
  aux did fit which does not make any sense)

This means that we should be now compliant with requirement stated in
Core 5.0, Vol 2, Part E, section 7.8.56:
    "The Controller should not start an extended advertising event
     that it cannot complete within the duration."

However, there's one tricky part here: we do not know whether full
extended advertising event can be sent within duration (i.e. with all
host data) as chain is scheduled on-the-fly using available air time.
For this reason we will just trim data if needed to complete extended
advertising event in time, as described earlier.