nimble/ll: Fix advertising interval validation
For legacy advertising any interval between 0x0020 and 0x4000 is valid
as long as min interval is not greater than max interval.

For extended advertising any interval between 0x000020 and 0xffffff
is valid as long as min interval is not greater than max interval.

If intervals provided are outside of value allowed by specificication
then Invalid Parmeters error is returned.

If the advertising interval range provided by the Host is valid but
outside the advertising interval range supported by the Controller,
then the Controller shall return the Unsupported Feature or Parameter
Value (0x11) error code. Since for now we support only intervals
in legacy range we use this code in that case.

This was affecting TP/DDI/BI-02-C qualification test case.