nimble/ll: Fix return from ble_ll_scan_start
This patch makes sure that ble_ll_scan_start return 0 (OK) if scanning
has been started or is ongoing.

It solves problem with incorrect free on scansm->cur_aux data when
ble_ll_scan_start return BLE_PHY_ERR_RX_LATE. In such a case radio was
scanning on secondary channel but scan state machine was not prepared
for it.