porting: Add FreeRTOS on nRF5 SDK porting example
nRF5 SDK (v15) is required to run this demo app. By default, it is
expected to be available at porting/examples/freertos_nrf5_sdk/nrf5_sdk
but this location can be changed by SDK_ROOT variable.

To build for nRF52832:
> cd porting/examples/freertos_nrf5_sdk
> make -C pca10040/blank/armgcc

To build for nRF52840:
> cd porting/examples/freertos_nrf5_sdk
> make -C pca10056/blank/armgcc

To build with nRF5 SDK placed in non-default location:
> cd porting/examples/freertos_nrf5_sdk
> make -C pca10040/blank/armgcc SDK_ROOT=/xxx/nRF5_SDK_15.0.0_a53641a

Flashing is done as for other examples in nRF5 SDK:
> make -C pca10040/blank/armgcc flash