nimble/phy: Fix late wfr setup
wfr is set using CC[0] or CC[2] as reference, but since CPU may be
halted for a brief moment between timer capture and wfr calculations it
may happen that TIMER0 already counted past calculated value and compare
will not occur as expected. This can leave radio stuck in RX state, e.g.
in case we are advertising and waiting for scan/connect req after TX.

To fix this we'll just check if wfr is set past current TIMER0 counter
value and if so, we'll disable radio manually as if wfr fired.

This is similar to 06cfc0b0 but happens for reverse transition.