nimble/host: Remove spurious conn-update callback
The GAP callback was incorrectly being called with an event type of
BLE_GAP_EVENT_CONN_UPDATE each time a connection was terminated.  For
example, a typical disconnect event might look like this in the log:

    023276 [ts=181843696us, mod=82 level=1] connection update; rc=531
    023277 [ts=181921816us, mod=82 level=1] disconnect; reason=531

In most cases, only the second line should be present.  The "connection
update" event should only be signaled on termination if an update
procedure for the connection happened to be in progress at the time.
In that case, the connection termination causes the update procedure to
fail.