nimble/ll: Fix legacy set adv data
When host use legacy HCI comands, it is possible (specification does not
forbid this) to set advertising data when directed advertiing is
configured.
This patch makes it possible as it fixes HCI/CM/BV-02 test when Tester
uses HCI_LE_Set_Advertising_Data to clear advertising data after it was
configured to directed advertising.