nimble/ll: Improve starting advertising for legacy pdu
In case advertising extensions are used for legacy advertising PDU let's
allow to start advertising even scan response data has not been set.
We do it like this as for legacy advertising it would force to set scan
response data as all connectable advertising is scannable then. We don't
want to do it.