host: Clear addresses on reset
A stack reset causes the controller to forget its random address.
Therefore, the host must reconfigure the controller with a random
address.

Prior to this commit, the host did not clear its addresses on stack
reset.  As a consequence, attempting to use a random address with
`ble_hs_util_ensure_addr()` function following a reset would not work.
The host would think that it had already configured the controller with
a random address, so this function would be a no-op.