host: Divide startup into two stages
Prior to this commit: host startup was done in a single stage by an
event on the default event queue.  This creates a problem if the
application wishes to run the host in a different event queue (i.e., not
the default task).  The typical init sequence looks like this:

1. Sysinit (start event enqueued to default event queue).
2. Application configures host, possible changing the host event queue.
3. Host start event runs in the default task.

The problem is that the start event in step three determines that the
host task is the currently running task.  This is always the default
task, which is incorrect if the user reconfigured the event queue.

Now: host startup is divided into two stages.  The first stage simply
enqueues a second event on the host event queue.  The startup sequence
now looks like this:

1. Sysinit (start event enqueued to default event queue).
2. Application configures host, possible changing the host event queue.
3. Stage 1 start event runs in the default task (stage 2 event enqueued
to host task).
4. Stage 2 start event runs in the host task.