Allow the host to be stopped and started again
The application can stop the host by calling `ble_hs_stop()`.  Stopping
the host has the following effects:
* Abort all active GAP procedures with status BLE_HS_EPREEMPTED.
* Disallow any subsequent GAP procedures until the host is started
  again.
* Terminate all established connections.

If multiple clients attempt to stop the host simultaneously, all of
their provided callbacks are executed when the host stops.  The purpose
of this behavior is to guarantee a consistent shutdown procedure.