nimble/ll: Improve handling disconnect when being Slave
When slave latency is greater than 0 we might not catch ACK on Terminate
Ind send by us being Slave.

At the moment, when Terminate Ind is sent, we keep listening for the ACK
however some devices waits with sending ACK for the next connection
event which we miss due to slave latency.

This patch fixes this.