apps/btshell: Add option to restart advertising after disconnect
This patch takes into account both legacy and extended advertising. When using
legacy advertising (BLE_EXT_ADV: 0) all advertising parameters are remembered
to use when restarting. When using extended advertising the parameters
per instance are remembered by the host. For ext adv the app remembers
the handle of the connection that issued Advertising Complete Event.
Later, at a Disconnect Event, the code checks if we should restart advertising
for an instance related to that connection.