nimble/ll: Request/release hfxo using mynewt API
This change is to use a new API committed to mynewt core that
will request the hfxo to be turned on and release it when the
hfxo is no longer desired. One reason to do this is that an
application may desire the hfxo to be on and nimble would
blindly turn it on/off when it wanted to. Note that the API
requires that the request/release be paired; you cannot release
without a request and if you request without a release the hfxo
will be on continuously.

A final note: the nimble link layer assumes that when it requests
the hfxo it needs to settle. If the hfxo was already on the code
would not need to wait for the settling time. The code has not
been updated to handle that case. This is not going to cause
any issues; it will operate as it did before. However, future
improvements could be to check if the hfxo was already on and
settled and thus not have to wait for the settling time.