nimble/ll: Extended scanner - add ref count for aux data
This patch improves scanner by adding ref counter for outstanding
aux_data. With this we want to make sure aux_data is not freed e.g. when
being handover to LL and in the same time processed in the interrupt in
chaining scenario

General idea is that 1 reference is kept as long as event is not
completed (note: truncated is also completed) and other references are
use when:
 - aux is scheduled
 - report is sending to the host
 - scan request has been sent
 - when packet is handover to LL