L2CAP sm: do not discard alignment information.
ble_sm_gen_ediv and ble_sm_gen_master_id_rand took a simple pointer to
integers, but were being called with pointers to members of a packed
struct. This discarded the alignment information.

Not only did it produce an error with clang's -Wno-address-of-packed-member,
but it may also cause an error in a platform with alignment requirements.

The solution was to modify both procedures to take a pointer to the whole
structure instead of only the members.