nimble/gatt: Fix the issue of removing multiple procs having same opcode
If two or more requests with same opcode are sent
(for e.g. Write and Enable/Disable notify), the requests go through and
responses are received. Although, the responses for all the requests
except for the first one are not sent above to the app (or upper
layers)
This is because, currently ble_gattc_extract_first_by_conn_op() extracts
all the matching procs and then they are removed from the list. Ideally,
it should extract the first matching proc (as its name suggests).

This patch modifies ble_gattc_extract_by_conn_op() which is used from
ble_gattc_extract_first_by_conn_op() to accept max_procs and extract the
appropriate number of procs, as required.