net/nimble/controller	Slave connection update fix
The code was not correctly handling the case where a slave receives
a connection update where the current connection event counter is
equal to the instant in the connection update PDU. This is a valid
case (as per spec errata). With this fix the controller should
correctly handle this case.