nimble/unittest: Adjust unittest for handling conn update parameters
Nimble changed the way of handling connection update parameter which
requires adjusting of unit tests. The main change is that from now on Nimble
will not timeout when it receive response on
L2CAP Connection Parameter Update Request at it is considered as the end of the
procedure from the Host point of view.