nimble/hw: Remove reading public address fom UICR
If public address was not present in FICR there was fallback to try to
read this address from UICR. However, how address is stored in UICR is
not standarized anywhere so it's just a vendor decision how and where
to put it. In worst case this may read some other data from UICR and
use them as public address in case vendor is not aware of this behavior.

To avoid such problems, let's remove this fallback code. If needed,
public address can now be read from virtually anywhere (also UICR) and
set using new API: ble_ll_set_public_addr(). This should be done in
e.g. BSP code and it's up to vendor.