nimble/ll: Fix update advertising data while advertising
With this patch we make sure that if advertising or scan data has been
changed just after advertising event had been completed, new data will
be used in next advertising event.

Note: For this moment we will skip one advertising event in order to
recalculate timings for aux packet, but this could be improved in the
future.