nimble/ll: Make sure that new DID is different than old one
Whenever the Host provides new advertising data or scan response data
for a given advertising set (whether it is the same as the previous
data or not), the Advertising DID shall be updated. The new value shall
be a randomly chosen value that is not the same as the previously used
value.

Signed-off-by: Szymon Janc <szymon.janc@codecoup.pl>