l2cap: Fix channel creation
Previously the channel was added to the channels list after sending
COC Accept event to the application. The application called
ble_l2cap_coc_recv_ready which was trying to find channel on a list,
but failed. With this patch the channel is added on a list before sending
Accept event, and if the application returns an error, the channel is removed
from the list. Moreover if we fail to send Connection Success, the application
should now not receive a confirmation that we are connected.