host tests: Don't use tu_suite_set_post_test_cb()
This testutil function was deprecated by this PR:
https://github.com/apache/mynewt-core/pull/1678

Instead, each test case needs to manually verify that all mbufs have
been freed.  This results in more code duplication, but automatically
performing this verification was not right anyway.  We do not want to
perform this step if a test case aborted early (failed
TEST_ASSERT_FATAL()).