nimble/sm: Fix bonding when key distribution is one-sided
When using Legacy pairing, Nimble is a slave/responder and
the initiator set key distribution to 0 then our LTK would
not be stored and it wouldn't be possible to restore encryption later.

In legacy mode, bonding requires the exchange of keys at least
from one side.  If no key exchange was specified, pretend bonding
is not enabled.

This fixes SM/SLA/KDU/BV-07-C.