host: Always wake tx queues on rx of num-cmp-pkts
Prior to commit: The code assumed that a transmission could stall for
only one reason: the controller cannot accommodate any more ACL data
packets.  When the host received a num-complered-packets event, it would
only try to unstall each connection if all the controller's data buffers
were previously used.

It turns out there is a second way that a host connection can stall:
mbuf exhaustion. If the host is unable to fragment an outgoing L2CAP
packet into ACL data packets due to a lack of mbufs, the transmission
stalls.

This commit causes the host to wake up its connections every time it
receives a num-completed-packets event. It does this even if the
controller still had capacity for more ACL data packets. If a host
connection is stalled due to mbuf exhaustion, the controller will only
send the next num-completed-packets event after it has freed the mbuf
from the designated ACL fragment pool (ble_hs_hci_frag_mbuf_pool). In
other words, we are guaranteed to receive a num-completed-packets event
whenever a transmission has stalled, regardless of the cause.