nimble/host: Allow NULL in Find attribute by UUID
Specifying UUID param as NULL will return any type of attribute.
This can be useful to get a list of attributes in the database.
For example it can be used in bttester application. It is an
internal API, so generally it should not be used by an application
but it is a testing application.