apps/btshell: Add time parsing function
New API allows to parse time strings which allow to specify values like
intervals, periods, timeouts etc. in an easy way instead of using units
which are always "I do not remember, need to check spec and calculate".

Time string is converted to microseconds and returned value is divided
by step value so it can be directly passed to various APIs.

Accepted time string format is "XXX.YYYuuu" where:
- XXX is integral part of value
- YYY is fractional part of value
- uuu is one of time units: "us", "ms" or "s"

XXX and YYY (and also separating ".") can be skipped and will be thus
set to 0.

An error is returned if format is invalid (in particular, if uuu is not
present) so this allows to fall back to standard integer conversion if
necessary.