nimble/ll: Update DID when enabling or disabling periodic advertising
The Advertising DID is not required to change when a SyncInfo field is
added to or removed from an advertising set. However, if it does not
change, then scanners may fail to synchronize to periodic advertising
because entries in the Advertising DID cache (see Section 4.3.3) mean
they ignore the advertisements containing the SyncInfo field. Therefore,
advertisers should update the Advertising DID when a periodic advertising
train is enabled. Alternatively, the Host should enable periodic
advertising before enabling advertising.