nimble/ll: Improve getting advertiser address for ext adv packet
With this change, when getting adva and inita from advertising
extensions, NimBLE will first try to get it from the packet and if not
succeed then it will look into related aux_data pointer.

This is needed especially when waiting for AUX_CONNECT_RSP where RPA
adva might be different from the one in advertising packets.