nimble/ll: Add ble_ll_scan_interrupt function
This is common fuction to be called whenever scanning has been
interrupted. This function sends event to LL task.

Also with this patch, scheduled AUX entry gets pointer to aux_data when
it is scheduled and not when aux_data is created. This is because
aux_data->ref_cnt is increased when scheduled.