nimble/mesh: Reject invalid remote public key before sending ours
Mesh Profile Spec v1.0.1 | Section 5.4.2.3:
`The Provisioner and the device shall check whether the public key
provided by the peer device or obtained OOB is valid (see Section 5.4.3.1).

When the Provisioner receives an invalid public key, then provisioning fails,
and the Provisioner shall act as described in Section 5.4.4. When the device
receives an invalid public key, then provisioning fails, and
the device shall act as described in Section 5.4.4.`

Described also in Erratum 10395 which is Mandatory for Mesh v1.0.

Fixes MESH/NODE/PROV/BI-13-C.

This test verifies that we do not send our public key if remote's
public key is invalid.