nimble/host: Simplify lookup for LE Meta event handler
Unlike HCI events, we have handler for almost all LE Meta events so it
is more efficient to have simple array of handlders indexed by subevent
code instead of having to iterate it as a list.

Also, since we do not need subevent code in dispatcher structure the new
array is smaller.