nimble: Use packed structures for HCI commands and events
This patch changes way HCI protocol is passed between host and
controller. Rule of thumb is that structures defined as packed
represent protocol order (LE) binary data, while normal structs
hold host order decoded (unpacked) data.

Packed structs are always used to parse or prepare HCI data. In
general HCI data is parsed only where needed and pass 'as-is'
up to stack as much as possible (this is to minimize need to copy
data).

Some additional changes and cleanups are also part of this commit
(due to invasive nature of this commit) to avoid unnecessary rebases
and conflicts:
 - constified pointers where needed (when passing packed data up)
 - updated HCI names as specified in Core 5.1

This commit not only improves code readiblity (ie it is easy to
check where data is used, as well as when it is expected to be
in host or in protocol order) but also reduces code size. Some
numbers for build with most features enabled (optimized build):
nimble_controller.a     55602 -> 55246
nimble_host.a           68001 -> 65285