nimble/ll: Do not use separate msys block for SCAN_REQ
LL allocates one mbuf on startup as a placeholder for SCAN_REQ. This
consumes one block from msys and was necessary to make sure we have
a buffer available instantly since SCAN_REQ is sent from interrupt
context.

Currently we can create PDUs on the fly during transmission so we only
need ScanA and AdvA to be calculated before starting TX and they will
be just copied directly to PDU while transmitting.

We cannot yet remove this extra block from state machine since it is
shared with CONNECT_IND, but we'll deal with this in next patch.