nimble/ll: Do not use separate msys buffer for CONNECT_IND
See previous commit.

We only need InitA and AdvA to create CONNECT_IND on the fly, no need
to have an extra always allocated mbuf. Storage for these data can be
shared with scan state machine, the same as we shared mbuf.

This allows to remove redundant mbuf allocation on startup and saves
one block of msys pool.