nimble/ll: Do not drop event on failed PDU reschedule
For unknown reason we drop advertising event when failed to reschedule
advertising PDU. This does not affect legacy advertising too much since
we'll just skip one or two channels, but in case of extended advertising
it also removes AUX which means complete extended advertising event is
lost.

In addition, nRF51 is not fast enough to ensure all PDUs on primary
channels can be scheduled before AUX so this effectively means extended
advertising on nRF51 is unusable.

With this fix we won't drop complete event if PDU reschedule fails, just
move to next channel. It allows extended advertising to be usable on
nRF51, but we should also allow to relax AUX scheduling constraints
which will allow to schedule all primary PDUs properly - this will be
addressed later.