nimble/ll: Fix error handling when sending chained adv report
In case there NimBLE has scheduled for next aux chain which is already
fired, but there was an issue with previous report sending over the HCI,
removing scheduled item from the scheduler will not succeed.

This patch makes sure truncated event will be delivered to the
host in such a case.