nimble/ll: Add support for custom DTM extensions
This allows to enable custom extensions for DTM which currently allow to
override interval between test packets and number of packets to be
transmitted. See description of new syscfg setting for details.

This is not enabled by default since it extends HCI command defined by
Core specification so we need to be compliant by default. However even
if extensions are enabled, NimBLE still accepts HCI command as defined
in Core specification but also accepts extended one.