nimble/ll: Fix receiving scan rsp with interleaved aux scan
This changes meaning of scan_rsp_pending in scan sm which now is set
only for duration of receiving scan response. When received packet
is passed to LL task information about it being SCAN_RSP is passed
in packet header flags. This allows for easy detection if scan
response is ongoing and bail out with aux scan that would interfere
with it.

This improves active scan backoff handling when extended advertising
is used.

This was affecting LL/DDI/SCN/BV-24-C and LL/DDI/SCN/BV-64-C
qualification test cases.