host/l2cap_coc: fix recv of frames with size 1
L2CAP payloads might be split into multiple frames, of which only
the first frame contains the 2-byte SDU length field and has
thereby a guaranteed length > 2. Any subsequent frame may be of
size 1 and are discarded without this fix. This commit fixes the
ble_l2cap_coc_rx_fn() so that it only calls
ble_hs_mbuf_pullup_base() if the incoming payload is >= 2 bytes.