nimble/ll: Round XTAL settle time up
This makes sure XTAL settle time is at least 1 tick, even if value is
specified as less than a full tick. A side-effect of having settle time
calculated as 0 ticks (when enabled) is that item which enables XTAL can
be run after scheduler item which needs it since they will be scheduled
on os_cputime queue at the same tick so it depends on order on which
they are put there.