nimble/transport/da1469x: Add CMAC driver
This adds CMAC driver which is responsible for initializing CMAC
(internal BLE controller) and loading Dialog's firmware.

This is (almost) the same code which was present in da1469x_cmac.c in
MCU package which will be removed from there. The reason is that driver
works with specific firmware which uses specific transport so it makes
sense to distribute them together. Also new driver is a bit smarter as
it does not need tricks to link Dialog's library with our code. Instead,
we just use object with firmware from that library to create our own
object with proper section names so this can play nicely with our linker
script (the firmware itself is not modified in any way).

Note that this requires recent newt version with support for build
scripts.